requirements.txt
Pillow==6.2.1
colorthief
tqdm
